Introduce new CLI module inspired by PicoCLI
The main idea of this implementation is to
determine all the necessary annotation metadata
at build time

This new module is used in the Quarkus CLI in
place of PicoCLI

+++ devtools/cli/src/test/java/io/quarkus/cli/image/CliImageGradleTest.java
@@ -20,20 +20,24 @@ import io.quarkus.cli.CliDriver;
 import io.quarkus.cli.common.build.ExecuteUtil;
 import io.quarkus.cli.common.build.GradleRunner;
 import io.quarkus.devtools.testing.RegistryClientTestHelper;
-import picocli.CommandLine;
+import io.quarkus.quickcli.ExitCode;
+import io.quarkus.test.junit.main.QuarkusMainLauncher;
+import io.quarkus.test.junit.main.QuarkusMainTest;
 
 /**
  * Similar to CliProjecGradleTest ..
  */
+@QuarkusMainTest
 public class CliImageGradleTest {
 
     static final Path testProjectRoot = Paths.get(System.getProperty("user.dir")).toAbsolutePath()
-            .resolve("target/test-classes/test-project/");
+            .resolve("target/test-project/");
     static final Path workspaceRoot = testProjectRoot.resolve("CliImageGradleTest");
     static final Path wrapperRoot = testProjectRoot.resolve("gradle-wrapper");
 
     Path project;
     static File gradle;
+    static boolean gradleDaemonStarted = false;
 
     @BeforeAll
     public static void setupTestRegistry() {
@@ -51,15 +55,18 @@ public class CliImageGradleTest {
         project = workspaceRoot.resolve("code-with-quarkus");
     }
 
-    @BeforeAll
-    static void startGradleDaemon() throws Exception {
+    static void ensureGradleDaemon(QuarkusMainLauncher launcher) throws Exception {
+        if (gradleDaemonStarted) {
+            return;
+        }
         CliDriver.deleteDir(wrapperRoot);
+        CliDriver.setLauncher(launcher);
 
         CliDriver.Result result = CliDriver.execute(workspaceRoot, "create", "app", "--gradle", "--verbose", "-e", "-B",
                 "--no-code",
                 "-o", testProjectRoot.toString(),
                 "gradle-wrapper");
-        Assertions.assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        Assertions.assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         Assertions.assertTrue(result.getStdout().contains("SUCCESS"),
                 "Expected confirmation that the project has been created." + result);
 
@@ -72,6 +79,7 @@ public class CliImageGradleTest {
 
         result = CliDriver.executeArbitraryCommand(wrapperRoot, args.toArray(new String[0]));
         Assertions.assertEquals(0, result.getExitCode(), "Gradle daemon should start properly");
+        gradleDaemonStarted = true;
     }
 
     @AfterAll
@@ -85,55 +93,58 @@ public class CliImageGradleTest {
             CliDriver.Result result = CliDriver.executeArbitraryCommand(wrapperRoot, args.toArray(new String[0]));
             Assertions.assertEquals(0, result.getExitCode(), "Gradle daemon should stop properly");
         }
+        gradleDaemonStarted = false;
     }
 
     @Test
-    public void testUsage() throws Exception {
+    public void testUsage(QuarkusMainLauncher launcher) throws Exception {
+        ensureGradleDaemon(launcher);
+        CliDriver.setLauncher(launcher);
         CliDriver.Result result = CliDriver.execute(workspaceRoot, "create", "app", "--gradle", "--verbose", "-e", "-B");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
 
         // 1 image --dry-run
         result = CliDriver.execute(project, "image", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertFalse(result.getStdout().contains("-Dquarkus.native.enabled=true"));
         result = CliDriver.execute(project, "image", "--native", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("-Dquarkus.native.enabled=true"));
 
         // 2 image build --dry-run
         result = CliDriver.execute(project, "image", "build", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=docker")); // Should fallback to docker
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "docker", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=docker"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "podman", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=podman"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "jib", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=jib"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "buildpack", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=buildpack"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "openshift", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("--builder=openshift"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         result = CliDriver.execute(project, "image", "build", "--group=mygroup", "--name=myname", "--tag=1.0", "--native",
                 "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.group=mygroup"));
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.name=myname"));
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.tag=1.0"));
@@ -141,13 +152,13 @@ public class CliImageGradleTest {
 
         // 3 image push --dry-run
         result = CliDriver.execute(project, "image", "push", "--dry-run");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.build=false"));
         assertTrue(result.getStdout().contains("--init-script="));
 
         // 4 image push --also-build --dry-run --registry=quay.io
         result = CliDriver.execute(project, "image", "push", "--also-build", "--dry-run", "--registry=quay.io");
-        assertEquals(CommandLine.ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
+        assertEquals(ExitCode.OK, result.getExitCode(), "Expected OK return code." + result);
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.build=true"));
         assertTrue(result.getStdout().contains("-Dquarkus.container-image.registry=quay.io"));
     }
